# wgengine: optimize isLocalAddr a bit